Clean the code up
