+++ fs/fs.c
@@ -274,40 +274,6 @@ int fs_ls(const char *dirname)
 
 	return ret;
 }
-/**
- * Omit a single arg and its value from envirment variable.
- * this single arg name should follow with white space, '\0' or '=' in case it has a value.
- * 
- *
- */
-int omitarg(const char *var_name, const char *arg_name){
-        char *var_val = getenv(var_name) ;
-
-        printf("debug: var_val = %s \n", var_val);
-        if(var_val==NULL){
-                printf("%s: no such envirument variable.\n", var_name );
-                return 0;
-        }
-        int arg_len= strlen(arg_name);
-
-        char *ptr= var_val ;
-        while( (  (ptr = strstr(ptr ,arg_name)) != NULL ) ){
-                if( ( ptr== var_val || *(ptr-1) == ' ' || *(ptr-1) == '\n' || *(ptr-1) == '\t' ||*(ptr-1) == '=' ) &&
-                        ( *(ptr+ arg_len)==' ' || *(ptr+arg_len)=='\n' || *(ptr+arg_len)=='\t' || *(ptr+arg_len)=='=' || *(ptr+arg_len)=='\0') ){
-                        char *arg_loc = ptr;
-                        while(*ptr!=' ' && *ptr!= '\t' && *ptr!= '\0'){
-                                *ptr=' ' ;
-                                ptr++;
-                        }
-                        strcpy( arg_loc , ptr  );
-                        return 1;
-                }
-                ptr = ptr + arg_len;
-        }
-        return 0;
-}
-
-
 
 int fs_exists(const char *filename)
 {
@@ -481,15 +447,6 @@ int do_ls(cmd_tbl_t *cmdtp, int flag, int argc, char * const argv[],
 		return 1;
 
 	return 0;
-}
-
-int do_omit(cmd_tbl_t *cmdtp, int flag, int argc, char * const argv[],
-        int fstype)
-{
-	printf("arg1 %s \n",  argv[1] ); 	
-	omitarg( argv[1], argv[2]); 
-	return 1 ; 
-
 }
 
 int file_exists(const char *dev_type, const char *dev_part, const char *file,
